+++ data_structures/clique.cc
