+++ old/connection.h
