+++ tools/checks.cc
