+++ tools/checks.h
